+++ src/index.ts
@@ -6,6 +6,7 @@ export { default as Button } from '@/package/components/button.vue';
 export { default as Input } from '@/package/components/input.vue';
 export { default as Textarea } from '@/package/components/textarea.vue';
 export { default as Select } from '@/package/components/select.vue';
+export { default as Autocomplete } from '@/package/components/autocomplete.vue';
 export { default as Checkbox } from '@/package/components/checkbox.vue';
 export { default as RadioGroup } from '@/package/components/radio-group.vue';
 export { default as RadioButton } from '@/package/components/radio-button.vue';